+++ lib/galaxy/webapps/galaxy/services/base.py
@@ -7,8 +7,7 @@ from typing import (
     Optional,
 )
 
-from celery.result import AsyncResult
-
+from galaxy.celery.helpers import async_task_summary as async_task_summary  # re-export for existing callers
 from galaxy.exceptions import (
     AuthenticationRequired,
     ConfigDoesNotAllowException,
@@ -32,7 +31,6 @@ from galaxy.model.store import (
 )
 from galaxy.schema.fields import EncodedDatabaseIdField
 from galaxy.schema.schema import (
-    AsyncTaskResultSummary,
     ToolRequestDetailedModel,
     ToolRequestModel,
 )
@@ -183,28 +181,6 @@ class ConsumesModelStores:
             history=history,
             for_library=for_library,
         )
-
-
-def async_task_summary(async_result: AsyncResult) -> AsyncTaskResultSummary:
-    name = None
-    try:
-        name = async_result.name
-    except AttributeError:
-        # if backend is disabled, we won't have this
-        pass
-    queue = None
-    try:
-        queue = async_result.queue
-    except AttributeError:
-        # if backend is disabled, we won't have this
-        pass
-
-    return AsyncTaskResultSummary(
-        id=str(async_result.id),
-        ignored=async_result.ignored,
-        name=name,
-        queue=queue,
-    )
 
 
 def _encode_tool_request(tool_request: ToolRequest, security: IdEncodingHelper) -> dict[str, Any]:
