+++ test/unit/authnz/test_psa_authnz.py
@@ -9,10 +9,7 @@ from datetime import (
 )
 from types import SimpleNamespace
 from typing import Optional
-from unittest.mock import (
-    MagicMock,
-    patch,
-)
+from unittest.mock import MagicMock
 
 import jwt
 import pytest
@@ -400,15 +397,16 @@ def test_oidc_config_custom_auth_pipeline_and_extra(mock_oidc_config_file, mock_
 def test_sync_user_profile_skips_when_account_interface_enabled():
     manager = MagicMock()
     session = MagicMock()
+    notify = MagicMock()
     app_config = SimpleNamespace(enable_account_interface=True, enable_notification_system=True)
-    app = SimpleNamespace(config=app_config, user_manager=manager, notification_manager=SimpleNamespace())
+    notification_manager = SimpleNamespace(send_notification_internal=notify)
+    app = SimpleNamespace(config=app_config, user_manager=manager, notification_manager=notification_manager)
     trans = SimpleNamespace(app=app, sa_session=session)
     strategy = SimpleNamespace(config={"GALAXY_TRANS": trans, "FIXED_DELEGATED_AUTH": True})
     user = SimpleNamespace(id=1, preferences={})
     details = {"email": "new@example.com", "username": "newname"}
 
-    with patch("galaxy.webapps.galaxy.services.notifications.NotificationService.send_notification_internal") as notify:
-        sync_user_profile(strategy=strategy, details=details, user=user)
+    sync_user_profile(strategy=strategy, details=details, user=user)
 
     manager.update_email.assert_not_called()
     manager.update_username.assert_not_called()
@@ -419,15 +417,16 @@ def test_sync_user_profile_skips_when_account_interface_enabled():
 def test_sync_user_profile_skips_when_fixed_delegated_auth_disabled():
     manager = MagicMock()
     session = MagicMock()
+    notify = MagicMock()
     app_config = SimpleNamespace(enable_account_interface=False, enable_notification_system=True)
-    app = SimpleNamespace(config=app_config, user_manager=manager, notification_manager=SimpleNamespace())
+    notification_manager = SimpleNamespace(send_notification_internal=notify)
+    app = SimpleNamespace(config=app_config, user_manager=manager, notification_manager=notification_manager)
     trans = SimpleNamespace(app=app, sa_session=session)
     strategy = SimpleNamespace(config={"GALAXY_TRANS": trans, "FIXED_DELEGATED_AUTH": False})
     user = SimpleNamespace(id=2, email="old@example.com", username="oldname", preferences={})
     details = {"email": "new@example.com", "username": "newname"}
 
-    with patch("galaxy.webapps.galaxy.services.notifications.NotificationService.send_notification_internal") as notify:
-        sync_user_profile(strategy=strategy, details=details, user=user)
+    sync_user_profile(strategy=strategy, details=details, user=user)
 
     manager.update_email.assert_not_called()
     manager.update_username.assert_not_called()
@@ -438,16 +437,16 @@ def test_sync_user_profile_skips_when_fixed_delegated_auth_disabled():
 def test_sync_user_profile_updates_when_account_interface_disabled():
     manager = MagicMock()
     session = MagicMock()
+    notify = MagicMock()
     app_config = SimpleNamespace(enable_account_interface=False, enable_notification_system=True)
-    notification_manager = SimpleNamespace(notifications_enabled=True)
+    notification_manager = SimpleNamespace(notifications_enabled=True, send_notification_internal=notify)
     app = SimpleNamespace(config=app_config, user_manager=manager, notification_manager=notification_manager)
     trans = SimpleNamespace(app=app, sa_session=session)
     strategy = SimpleNamespace(config={"GALAXY_TRANS": trans, "FIXED_DELEGATED_AUTH": True})
     user = SimpleNamespace(id=2, email="old@example.com", username="oldname", preferences={})
     details = {"email": "new@example.com", "username": "newname"}
 
-    with patch("galaxy.webapps.galaxy.services.notifications.NotificationService.send_notification_internal") as notify:
-        sync_user_profile(strategy=strategy, details=details, user=user)
+    sync_user_profile(strategy=strategy, details=details, user=user)
 
     manager.update_email.assert_called_once_with(
         trans, user, "new@example.com", commit=False, send_activation_email=False
